requirements.txt
fastapi
uvicorn
jinja2
python-multipart
numpy
opencv-python-headless
tensorflow-cpu